+++ pdi/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### For users
 
 #### Added
+* Improved error messages when failing to load a plugin
 
 #### Changed